+++ source/blender/draw/engines/overlay/shaders/overlay_edit_curve_point_vert.glsl
@@ -7,7 +7,7 @@ void main()
   GPU_INTEL_VERTEX_SHADER_WORKAROUND
 
   /* Reuse the FREESTYLE flag to determine is GPencil. */
-  bool is_gpencil = ((data & EDGE_FREESTYLE) != 0);
+  bool is_gpencil = ((data & EDGE_FREESTYLE) != 0u);
   if ((data & VERT_SELECTED) != 0u) {
     if ((data & VERT_ACTIVE) != 0u) {
       finalColor = colorEditMeshActive;
@@ -26,7 +26,7 @@ void main()
   view_clipping_distances(world_pos);
 
   bool show_handle = showCurveHandles;
-  if ((curveHandleDisplay == CURVE_HANDLE_SELECTED) && ((data & VERT_SELECTED_BEZT_HANDLE) == 0u)) {
+  if (uint(curveHandleDisplay == CURVE_HANDLE_SELECTED) && ((data & VERT_SELECTED_BEZT_HANDLE) == 0u)) {
     show_handle = false;
   }
